Synthesizers and Vocoders added (but they does not sound nice)

+++ src/wavbuffer/WavBuffer.java
@@ -34,7 +34,7 @@ public class WavBuffer<C> {
         return this.right;
     }
 
-    public int size() {
+    public long size() {
         return this.left.size();
     }
